backend/app/soar/requirements.txt
requests
pandas
python-dotenv
colorama
fpdf2
